Spec Changes and Schematic
Tiny bit of progress on the schematic and change to the feature set. The
sub 1GHZ radio is going to be integrated into the same board as the
beagle sip, I'm doing away with the USB hub for now. Doing away with the
wireless chip from Ti going to use a ESP32 as a realtime controller and
for standard wireless since the focus is remote wireless not indoor
wireless. Need to fit all the important stuff first before adding extra
features like a USB hub.